Fix #122606: fully map Blender's INT8 type to USD's UChar
While USD does not have a signed, 8-bit, integer type we can use its
UChar type instead.

Pull Request: https://projects.blender.org/blender/blender/pulls/122610

+++ tests/python/bl_usd_import_test.py
@@ -705,7 +705,7 @@ class USDImportTest(AbstractUSDTest):
         mesh = bpy.data.objects["Mesh"].data
 
         self.check_attribute(mesh, "p_bool", 'POINT', 'BOOLEAN', 4)
-        self.check_attribute(mesh, "p_int8", 'POINT', 'INT', 4)
+        self.check_attribute(mesh, "p_int8", 'POINT', 'INT8', 4)
         self.check_attribute(mesh, "p_int32", 'POINT', 'INT', 4)
         self.check_attribute(mesh, "p_float", 'POINT', 'FLOAT', 4)
         self.check_attribute(mesh, "p_byte_color", 'POINT', 'FLOAT_COLOR', 4)
@@ -716,7 +716,7 @@ class USDImportTest(AbstractUSDTest):
         self.check_attribute_missing(mesh, "p_mat4x4")
 
         self.check_attribute(mesh, "f_bool", 'FACE', 'BOOLEAN', 1)
-        self.check_attribute(mesh, "f_int8", 'FACE', 'INT', 1)
+        self.check_attribute(mesh, "f_int8", 'FACE', 'INT8', 1)
         self.check_attribute(mesh, "f_int32", 'FACE', 'INT', 1)
         self.check_attribute(mesh, "f_float", 'FACE', 'FLOAT', 1)
         self.check_attribute_missing(mesh, "f_byte_color")  # Not supported?
@@ -727,7 +727,7 @@ class USDImportTest(AbstractUSDTest):
         self.check_attribute_missing(mesh, "f_mat4x4")
 
         self.check_attribute(mesh, "fc_bool", 'CORNER', 'BOOLEAN', 4)
-        self.check_attribute(mesh, "fc_int8", 'CORNER', 'INT', 4)
+        self.check_attribute(mesh, "fc_int8", 'CORNER', 'INT8', 4)
         self.check_attribute(mesh, "fc_int32", 'CORNER', 'INT', 4)
         self.check_attribute(mesh, "fc_float", 'CORNER', 'FLOAT', 4)
         self.check_attribute(mesh, "fc_byte_color", 'CORNER', 'FLOAT_COLOR', 4)
@@ -743,7 +743,7 @@ class USDImportTest(AbstractUSDTest):
         curves = curves[0].data
 
         self.check_attribute(curves, "p_bool", 'POINT', 'BOOLEAN', 24)
-        self.check_attribute(curves, "p_int8", 'POINT', 'INT', 24)
+        self.check_attribute(curves, "p_int8", 'POINT', 'INT8', 24)
         self.check_attribute(curves, "p_int32", 'POINT', 'INT', 24)
         self.check_attribute(curves, "p_float", 'POINT', 'FLOAT', 24)
         self.check_attribute_missing(curves, "p_byte_color")
@@ -754,7 +754,7 @@ class USDImportTest(AbstractUSDTest):
         self.check_attribute_missing(curves, "p_mat4x4")
 
         self.check_attribute(curves, "sp_bool", 'CURVE', 'BOOLEAN', 2)
-        self.check_attribute(curves, "sp_int8", 'CURVE', 'INT', 2)
+        self.check_attribute(curves, "sp_int8", 'CURVE', 'INT8', 2)
         self.check_attribute(curves, "sp_int32", 'CURVE', 'INT', 2)
         self.check_attribute(curves, "sp_float", 'CURVE', 'FLOAT', 2)
         self.check_attribute_missing(curves, "sp_byte_color")
@@ -769,7 +769,7 @@ class USDImportTest(AbstractUSDTest):
         curves = curves[0].data
 
         self.check_attribute(curves, "p_bool", 'POINT', 'BOOLEAN', 10)
-        self.check_attribute(curves, "p_int8", 'POINT', 'INT', 10)
+        self.check_attribute(curves, "p_int8", 'POINT', 'INT8', 10)
         self.check_attribute(curves, "p_int32", 'POINT', 'INT', 10)
         self.check_attribute(curves, "p_float", 'POINT', 'FLOAT', 10)
         self.check_attribute_missing(curves, "p_byte_color")
@@ -780,7 +780,7 @@ class USDImportTest(AbstractUSDTest):
         self.check_attribute_missing(curves, "p_mat4x4")
 
         self.check_attribute(curves, "sp_bool", 'CURVE', 'BOOLEAN', 3)
-        self.check_attribute(curves, "sp_int8", 'CURVE', 'INT', 3)
+        self.check_attribute(curves, "sp_int8", 'CURVE', 'INT8', 3)
         self.check_attribute(curves, "sp_int32", 'CURVE', 'INT', 3)
         self.check_attribute(curves, "sp_float", 'CURVE', 'FLOAT', 3)
         self.check_attribute_missing(curves, "sp_byte_color")
